Extracted test_suite_instances
